Break into several files, add tests
